🐛 Separated codes to fix import order error
