+++ src-electron/source/version/version.test.ts
@@ -1,0 +1,148 @@
+import { describe, expect, test, vi } from 'vitest';
+import { Failable } from 'app/src-electron/schema/error';
+import { Path } from 'app/src-electron/util/binary/path';
+import { isError } from 'app/src-electron/util/error/error';
+import { VersionContainer } from './version';
+
+describe('listupVersions', async () => {
+  // 一時使用フォルダを初期化
+  const workPath = new Path(__dirname).child('getVersions/work/versions');
+  await workPath.emptyDir();
+
+  const vc = new VersionContainer(workPath);
+
+  // キャッシュの威力を確認するときにはTrueにする
+  const useCache = true;
+
+  type TestCase = {
+    vcList: () => Promise<Failable<any>>;
+    oldestVersion: string;
+  };
+  const testCases: TestCase[] = [
+    {
+      vcList: () => vc.listVersions('vanilla', useCache),
+      oldestVersion: '1.3',
+    },
+    {
+      vcList: () => vc.listVersions('forge', useCache),
+      oldestVersion: '1.5.2',
+    },
+    {
+      vcList: () => vc.listVersions('papermc', useCache),
+      oldestVersion: '1.8.8',
+    },
+    {
+      vcList: () => vc.listVersions('mohistmc', useCache),
+      oldestVersion: '1.7.10',
+    },
+    {
+      vcList: () => vc.listVersions('fabric', useCache),
+      oldestVersion: '18w43b',
+    },
+    {
+      vcList: () => vc.listVersions('spigot', useCache),
+      oldestVersion: '1.9',
+    },
+  ];
+
+  test.each(testCases)('versionList ($type)', async (tCase) => {
+    const getCachedList = await tCase.vcList(); // 取得に成功したか
+    expect(isError(getCachedList)).toEqual(false);
+    if (isError(getCachedList)) return;
+
+    // 取得した内容が正しいか（バニラの最も古いバージョンは「1.3」）
+    const cachedList = getCachedList;
+    if ('games' in cachedList) {
+      // fabricのみ特別対応
+      expect(cachedList.games[cachedList.games.length - 1].id).toEqual(
+        tCase.oldestVersion
+      );
+    } else {
+      // その他のサーバー
+      expect(cachedList[cachedList.length - 1].id).toEqual(tCase.oldestVersion);
+    }
+  });
+});
+
+describe(
+  'eula生成テスト',
+  async () => {
+    const { VanillaVersion } = await import('../../schema/version');
+
+    const workPath = new Path(__dirname).child('work', 'version');
+    await workPath.remove();
+    const container = new VersionContainer(workPath.child('cache'));
+
+    const vanilla = VanillaVersion.parse({
+      type: 'vanilla',
+      id: '1.12.1',
+      release: false,
+    });
+
+    const serverPath = workPath.child('server');
+
+    // 1回目 (eula=false)
+    test('1回目 (eula=false)', async () => {
+      const eulaDisagree = vi.fn(async () => false);
+      const readyResult = await container.readyVersion(
+        vanilla,
+        serverPath,
+        async () => {},
+        eulaDisagree
+      );
+      // eulaの同意を求める関数が実行されるはず
+      expect(eulaDisagree).toHaveBeenCalledTimes(1);
+      // eula.txtが "eula=false" であるはず
+      expect(await serverPath.child('eula.txt').readText()).toBe('eula=false');
+      // readyResultが失敗するはず
+      expect(isError(readyResult)).toBe(true);
+
+      // 撤収
+      await container.removeVersion(vanilla, serverPath);
+      expect(serverPath.child('eula.txt').exists()).toBe(false);
+    });
+
+    // 2回目 (eula=true)
+    test('2回目 (eula=true)', async () => {
+      const eulaAgree = vi.fn(async () => true);
+      const readyResult = await container.readyVersion(
+        vanilla,
+        serverPath,
+        async () => {},
+        eulaAgree
+      );
+      // eulaの同意を求める関数が実行されるはず
+      expect(eulaAgree).toHaveBeenCalledTimes(1);
+      // eula.txtが "eula=true" であるはず
+      expect(await serverPath.child('eula.txt').readText()).toBe('eula=true');
+      // readyResultが成功するはず
+      expect(isError(readyResult)).toBe(false);
+
+      // 撤収
+      await container.removeVersion(vanilla, serverPath);
+      expect(serverPath.child('eula.txt').exists()).toBe(false);
+    });
+
+    // 3回目 (eula=true)
+    test('3回目 (eula=true)', async () => {
+      const eulaAgree = vi.fn(async () => true);
+      const readyResult = await container.readyVersion(
+        vanilla,
+        serverPath,
+        async () => {},
+        eulaAgree
+      );
+      // eulaの同意を求める関数は既に同意済みなので実行されないはず
+      expect(eulaAgree).toHaveBeenCalledTimes(0);
+      // eula.txtが "eula=true" であるはず
+      expect(await serverPath.child('eula.txt').readText()).toBe('eula=true');
+      // readyResultが成功するはず
+      expect(isError(readyResult)).toBe(true);
+
+      // 撤収
+      await container.removeVersion(vanilla, serverPath);
+      expect(serverPath.child('eula.txt').exists()).toBe(false);
+    });
+  },
+  1000 * 60
+);
