+++ src-electron/source/version/versionLogger.ts
@@ -1,0 +1,3 @@
+import { sourceLoggers } from '../sourceLogger';
+
+export const versionLogger = () => sourceLoggers().version;
